Phase 4.5: Replace old codegen with new implementation
- Delete old codegen.rs (imperative style with CodegenState)
- Rename codegen_judgment → codegen
- Remove dead grammar/expr/codegen.rs (abandoned earlier attempt)
- All 29 codegen tests pass, full test suite green
